Removed legacy component
