+++ sparkle/src/components/markdown/index.ts
@@ -1,6 +1,5 @@
 export * from "./CodeBlock";
 export * from "./CodeBlockWithExtendedSupport";
 export * from "./ContentBlockWrapper";
-export * from "./ExtendedMarkdown";
-export * from "./LegacyMarkdown";
+export * from "./Markdown";
 export * from "./MarkdownContentContext";
